solver: add support for PrA solution procedure
This patch adds an implementation of the basic Presburger
Arithmetic solver. The procedure is based on a classical automata-based
approach for proving Presburger arithmetic statements.

For each atomic formula an automaton is built and then they're
intersected/unioned corresponding to the atomic formula logical
connectives.

Now, when running the solver, the REPL is provided, supporting a variety
of basic commands:
* `eval` for proving/disproving statements.
* `parse` for dumping the AST for the formula.
* `dump` for drawing the NFA using GraphViz.

Closes #12