solver: fix small statement solution
This patch fixes problems that occur when solving simple statements with
the same amount of internal variables. E.g. `2 <= 3`.

The problem was in the fact the NFAs for their values used the same
internal variable indices in the vectors over the labels. I.e. in
case of `2 <= 3` it's both `0` thus the intersection of the automata
gave wrong results.

This patch fixes it by making the solver reset the internal counter after
solving the atomic formula.