solver: add support for predicates
This patch adds support for predicates that's been reverted in fda25233
('nfa: drastically speed-up basic operations') due to migration to
bitvectors for representing edge labels. The implementation is similar
to the previous. The bitvectors are re-enumerated correspondingly to the
variable names.

Predicates can be used as follows.
```
> let even x = Ey x = 2y

> eval even(2)
Result: true

> eval Ax even(x) -> ~even(x + 1)
Result: true
```