solver: support getting a model for PrA formula
This patch allows user to find a model for free variables of a formula
by running `evalm formula`.

Example.

```
> evalm (Ex y = 15x) & y > 64 & y < 128
  y = 75
```

Basically, the model is sought by evaluating deep-first traversal of
the resulting automaton from one of the starting states to one of the
final ones and then collecting the labels into integer values.

Semёnov arithmetic support will be added later.