solver: speedup Semenov arithmetic
This patch drastically speedups Semenov arithmetic evaluation. Now the
variables that aren't used within the `2**` functional symbol are
automatically projected and ignored while trying different orders.

This means it's a lot less orders in general to try to.

It also contains some small fix ups: `and` and `or` in smtlib can now
contain arbitrary amount of arguments.

Also, the parser now properly handles open parens when in formulas.